requirements.txt
absl-py==0.9.0
astor==0.8.1
autoflake==1.3.1
autopep8==1.5.1
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
gast==0.2.2
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
graphviz==0.13.2
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.0
pandas==1.0.3
parso==0.6.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.2
subprocess32==3.5.4
tensorboard==2.1.1
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
wandb==0.8.32
watchdog==0.10.2
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
